uiとmodelviewの分離
